2.0.4; General Refactoring and Updates
Bump versions of dependencies, see the 'build.gradle.kts'.
Bump version to Java 21.
Bump Gradle to version 8.5.

Support Minecraft 1.20.6.
  - Now uses glint override to make an item glow in ItemCreator, this class is planned for an overhaul.

Support to set a menu Button to multiple positions.
Add default manager for inventories, this is likely to be removed in the future in favor of a better system.
The IMenuManager interface now contains 'remove' methods.
Added utility methods to obtain an InventoryPosition based on a slot in an inventory (any integer).
ItemCreator now supports setting custom model data.
Exposes reschedulable menu tasks.
We now make a new ArrayList for the page items.
Added method to refresh a viewed page in a PageMenu.

Added ADDITIONAL_RESOLVERS variable to StringUtils to allow registering custom MiniMessage tags that can be used in all StringUtil.color calls.
Moved string wrap utils to it's own class and improved it.

Added additional utility methods to SubCommandBase.

Use the Bukkit logger instead of sending a message to the ConsoleCommandSender.
  - This comes with additional methods for logging warnings and proper support for error logging.