2.0.5; Add Brigadier command support.
Update paperweight to 1.7.1.
Update paper to 1.21 snapshots.
Removed JetBrains annotations from the compiled jar.

Use regex for replacing all & or § in color strings.

Use Utils#logError to log more errors in normal commands and in SimpleConfig.
Implement methods in PlayerCommand and PlayerSubCommand to immediately assume the completer is a player.

Minor corrections to some documentation and removal of some comments that aren't really needed.