Update to 1.21.5
AnvilGUI now uses a more modern way of opening an anvil menu.
Marked AnvilGUI's InventoryListener as 'final.'

Removed unnecessary TODO in common/build.gradle.kts; as it now functions properly.
GenericEnumArgument will now properly filter arguments based on the input.
Fixed an issue with CommandRequirements where the PLAYER and CONSOLE predicates would fail.
Updated ISkull to require a UUID instead of a name; this is to enforce the idea that UUIDs are a better match for Textures instead of names.
Fixed a bug with ItemCreator lore not throwing NullPointer because the list was never set.
Reworked ItemOptions a tiny bit to use the TooltipDisplay components for hiding things.
Deprecated ItemOptions.HIDE_ALL_FLAGS as you should pass the components you wish to hide instead of all components, this is planned for removal in 2.1.3.
SkinTexture now substitutes names with an empty string.
Weakened visibility for methods in AlternateResolver.
Added StringUtil#toFormattedList to convert a collection of something into a formatted string separated by another string and replacing the last iteration of the separator with 'and.'
Utils can now take a Logger so logging can be performed with a bootloader or at any point where a plugin instance cannot be properly registered.
Updated the ItemStack serialization to base64 methods to use built in Paper methods.
Prebuilt buttons no longer hide all tags by default.
Added 'canReturn' functionality to ReturnButton, allowing for checks to be conducted before allowing the return to a previous menu.
Fix an issue with ListMenu where it wouldn't properly draw the list buttons.
Fixed some inconsistencies in the PaginatedMenu.
MenuFunction added.

Minor updates to Wand method names, this does break previous wands.
Updated the maps that store locations to use the final static naming conventions; similarly updated the Listener variable, despite it not being static.

Signed-off-by: ItsSchatten <itsschatten@itsschatten.com>